adding attribute `ogit/access-control`
This initial discussion of this attribute was for it to be `ogit/public` meaning
public read-only access to the node. However on creating this pull-request I
realised that abstracting this into an ACL value would provide applications more
flexibility than a binary `public`/`not public`.

I did struggle with the node name, `access-control-list` seemed execessively
verbose, `acl` seemed like too much of a contraction/intialism, and `access`
seemed too vague.